Encrypt a text file. Version 1
Basis to encrypt a text file. The code must be improved. Function added to
read a dictionary file and create a dictionary (as an object) used for
encrypting a text file.

/!\ BUG !!!! I have found a BUG with the function assign_key: pointer dict->key is
dynamic so it is modified when reading a new line in function
upload_dict()... At the moment it does not prevent us from encrypting a
file.

BUG SOLVED : unable to use 'strnlen' because this function is not
implemented in my header string.h ...